release 3.1.42 source code for java

+++ services/bss/src/main/java/com/huaweicloud/sdk/bss/v2/model/AccountBalanceV3.java
@@ -3,6 +3,7 @@ package com.huaweicloud.sdk.bss.v2.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
 import java.util.Objects;
 
 /**
@@ -23,7 +24,7 @@ public class AccountBalanceV3 {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "amount")
 
-    private Double amount;
+    private BigDecimal amount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "currency")
@@ -33,12 +34,12 @@ public class AccountBalanceV3 {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "designated_amount")
 
-    private Double designatedAmount;
+    private BigDecimal designatedAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "credit_amount")
 
-    private Double creditAmount;
+    private BigDecimal creditAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "measure_id")
@@ -79,7 +80,7 @@ public class AccountBalanceV3 {
         this.accountType = accountType;
     }
 
-    public AccountBalanceV3 withAmount(Double amount) {
+    public AccountBalanceV3 withAmount(BigDecimal amount) {
         this.amount = amount;
         return this;
     }
@@ -88,11 +89,11 @@ public class AccountBalanceV3 {
      * 账户余额。
      * @return amount
      */
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
 
@@ -113,7 +114,7 @@ public class AccountBalanceV3 {
         this.currency = currency;
     }
 
-    public AccountBalanceV3 withDesignatedAmount(Double designatedAmount) {
+    public AccountBalanceV3 withDesignatedAmount(BigDecimal designatedAmount) {
         this.designatedAmount = designatedAmount;
         return this;
     }
@@ -122,15 +123,15 @@ public class AccountBalanceV3 {
      * 专款专用余额。
      * @return designatedAmount
      */
-    public Double getDesignatedAmount() {
+    public BigDecimal getDesignatedAmount() {
         return designatedAmount;
     }
 
-    public void setDesignatedAmount(Double designatedAmount) {
+    public void setDesignatedAmount(BigDecimal designatedAmount) {
         this.designatedAmount = designatedAmount;
     }
 
-    public AccountBalanceV3 withCreditAmount(Double creditAmount) {
+    public AccountBalanceV3 withCreditAmount(BigDecimal creditAmount) {
         this.creditAmount = creditAmount;
         return this;
     }
@@ -139,11 +140,11 @@ public class AccountBalanceV3 {
      * 总信用额度，仅信用账户存在该字段。
      * @return creditAmount
      */
-    public Double getCreditAmount() {
+    public BigDecimal getCreditAmount() {
         return creditAmount;
     }
 
-    public void setCreditAmount(Double creditAmount) {
+    public void setCreditAmount(BigDecimal creditAmount) {
         this.creditAmount = creditAmount;
     }
